Added checks on context and disposing of UserPrincipal objects
Renamed  currentDomainContext to currentDomainContextName to prevent confusing it as a PrincipalContext when it is just the string name of the context pulled from the ComboBox.

Added check that credentials exist for the configured domain in the drop down box. If it does not, it defaults back to the currently connected domain.

Next steps will be making sure a user is currently connected to a domain. Such as, they forgot to connect to VPN before running the application / etc.